+++ geocoding/dev/src/payload-types.ts
@@ -69,6 +69,7 @@ export interface Config {
   collections: {
     users: User;
     pages: Page;
+    articles: Article;
     'payload-kv': PayloadKv;
     'payload-locked-documents': PayloadLockedDocument;
     'payload-preferences': PayloadPreference;
@@ -78,13 +79,14 @@ export interface Config {
   collectionsSelect: {
     users: UsersSelect<false> | UsersSelect<true>;
     pages: PagesSelect<false> | PagesSelect<true>;
+    articles: ArticlesSelect<false> | ArticlesSelect<true>;
     'payload-kv': PayloadKvSelect<false> | PayloadKvSelect<true>;
     'payload-locked-documents': PayloadLockedDocumentsSelect<false> | PayloadLockedDocumentsSelect<true>;
     'payload-preferences': PayloadPreferencesSelect<false> | PayloadPreferencesSelect<true>;
     'payload-migrations': PayloadMigrationsSelect<false> | PayloadMigrationsSelect<true>;
   };
   db: {
-    defaultIDType: string;
+    defaultIDType: number;
   };
   fallbackLocale: null;
   globals: {};
@@ -122,7 +124,7 @@ export interface UserAuthOperations {
  * via the `definition` "users".
  */
 export interface User {
-  id: string;
+  id: number;
   updatedAt: string;
   createdAt: string;
   email: string;
@@ -147,110 +149,96 @@ export interface User {
  * via the `definition` "pages".
  */
 export interface Page {
-  id: string;
+  id: number;
   title?: string | null;
-  location_googlePlacesData?:
-    | {
-        [k: string]: unknown;
-      }
-    | unknown[]
-    | string
-    | number
-    | boolean
-    | null;
+  location_meta?: {
+    name: string;
+    formattedAddress: string;
+    googlePlaceId: string;
+    types: string[];
+  };
   /**
    * @minItems 2
    * @maxItems 2
    */
   location?: [number, number] | null;
-  location0_googlePlacesData?:
-    | {
-        [k: string]: unknown;
-      }
-    | unknown[]
-    | string
-    | number
-    | boolean
-    | null;
+  location_address?: string | null;
+  location0_meta?: {
+    name: string;
+    formattedAddress: string;
+    googlePlaceId: string;
+    types: string[];
+  };
   /**
    * @minItems 2
    * @maxItems 2
    */
   location0?: [number, number] | null;
-  location1_googlePlacesData?:
-    | {
-        [k: string]: unknown;
-      }
-    | unknown[]
-    | string
-    | number
-    | boolean
-    | null;
+  location0_address?: string | null;
+  location1_meta?: {
+    name: string;
+    formattedAddress: string;
+    googlePlaceId: string;
+    types: string[];
+  };
   /**
    * @minItems 2
    * @maxItems 2
    */
   location1: [number, number];
-  location2_googlePlacesData:
-    | {
-        [k: string]: unknown;
-      }
-    | unknown[]
-    | string
-    | number
-    | boolean
-    | null;
+  location1_address?: string | null;
+  location2_meta: {
+    name: string;
+    formattedAddress: string;
+    googlePlaceId: string;
+    types: string[];
+  };
   /**
    * @minItems 2
    * @maxItems 2
    */
   location2: [number, number];
-  location3_googlePlacesData?:
-    | {
-        [k: string]: unknown;
-      }
-    | unknown[]
-    | string
-    | number
-    | boolean
-    | null;
+  location2_address?: string | null;
+  location3_meta?: {
+    name: string;
+    formattedAddress: string;
+    googlePlaceId: string;
+    types: string[];
+  };
   /**
    * @minItems 2
    * @maxItems 2
    */
   location3?: [number, number] | null;
+  location3_address?: string | null;
   locationGroup?: {
-    location_googlePlacesData?:
-      | {
-          [k: string]: unknown;
-        }
-      | unknown[]
-      | string
-      | number
-      | boolean
-      | null;
+    location_meta?: {
+      name: string;
+      formattedAddress: string;
+      googlePlaceId: string;
+      types: string[];
+    };
     /**
      * @minItems 2
      * @maxItems 2
      */
     location?: [number, number] | null;
+    location_address?: string | null;
   };
   locations?:
     | {
-        location_googlePlacesData?:
-          | {
-              [k: string]: unknown;
-            }
-          | unknown[]
-          | string
-          | number
-          | boolean
-          | null;
+        location_meta?: {
+          name: string;
+          formattedAddress: string;
+          googlePlaceId: string;
+          types: string[];
+        };
         /**
          * @minItems 2
          * @maxItems 2
          */
         location?: [number, number] | null;
+        location_address?: string | null;
         id?: string | null;
       }[]
     | null;
@@ -259,10 +247,35 @@ export interface Page {
 }
 /**
  * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "articles".
+ */
+export interface Article {
+  id: number;
+  title: string;
+  content?: {
+    root: {
+      type: string;
+      children: {
+        type: any;
+        version: number;
+        [k: string]: unknown;
+      }[];
+      direction: ('ltr' | 'rtl') | null;
+      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
+      indent: number;
+      version: number;
+    };
+    [k: string]: unknown;
+  } | null;
+  updatedAt: string;
+  createdAt: string;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
  * via the `definition` "payload-kv".
  */
 export interface PayloadKv {
-  id: string;
+  id: number;
   key: string;
   data:
     | {
@@ -279,20 +292,24 @@ export interface PayloadKv {
  * via the `definition` "payload-locked-documents".
  */
 export interface PayloadLockedDocument {
-  id: string;
+  id: number;
   document?:
     | ({
         relationTo: 'users';
-        value: string | User;
+        value: number | User;
       } | null)
     | ({
         relationTo: 'pages';
-        value: string | Page;
+        value: number | Page;
+      } | null)
+    | ({
+        relationTo: 'articles';
+        value: number | Article;
       } | null);
   globalSlug?: string | null;
   user: {
     relationTo: 'users';
-    value: string | User;
+    value: number | User;
   };
   updatedAt: string;
   createdAt: string;
@@ -302,10 +319,10 @@ export interface PayloadLockedDocument {
  * via the `definition` "payload-preferences".
  */
 export interface PayloadPreference {
-  id: string;
+  id: number;
   user: {
     relationTo: 'users';
-    value: string | User;
+    value: number | User;
   };
   key?: string | null;
   value?:
@@ -325,7 +342,7 @@ export interface PayloadPreference {
  * via the `definition` "payload-migrations".
  */
 export interface PayloadMigration {
-  id: string;
+  id: number;
   name?: string | null;
   batch?: number | null;
   updatedAt: string;
@@ -359,29 +376,46 @@ export interface UsersSelect<T extends boolean = true> {
  */
 export interface PagesSelect<T extends boolean = true> {
   title?: T;
-  location_googlePlacesData?: T;
+  location_meta?: T;
   location?: T;
-  location0_googlePlacesData?: T;
+  location_address?: T;
+  location0_meta?: T;
   location0?: T;
-  location1_googlePlacesData?: T;
+  location0_address?: T;
+  location1_meta?: T;
   location1?: T;
-  location2_googlePlacesData?: T;
+  location1_address?: T;
+  location2_meta?: T;
   location2?: T;
-  location3_googlePlacesData?: T;
+  location2_address?: T;
+  location3_meta?: T;
   location3?: T;
+  location3_address?: T;
   locationGroup?:
     | T
     | {
-        location_googlePlacesData?: T;
+        location_meta?: T;
         location?: T;
+        location_address?: T;
       };
   locations?:
     | T
     | {
-        location_googlePlacesData?: T;
+        location_meta?: T;
         location?: T;
+        location_address?: T;
         id?: T;
       };
+  updatedAt?: T;
+  createdAt?: T;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "articles_select".
+ */
+export interface ArticlesSelect<T extends boolean = true> {
+  title?: T;
+  content?: T;
   updatedAt?: T;
   createdAt?: T;
 }